% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautify_df.R
\name{beautify_df}
\alias{beautify_df}
\title{Master Function to Beautify a DataFrame}
\usage{
beautify_df(df, highlight = c("na", "max", "min"), colors = list())
}
\arguments{
\item{df}{A dataframe.}

\item{highlight}{Choose which to highlight: "na", "max", "min", or multiple.}

\item{colors}{List of colors for each highlight.}
}
\value{
A beautified datatable.
}
\description{
Master Function to Beautify a DataFrame
}
