% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{main function to run linear regression with mini bag bootstrap}
\usage{
blblm(formula, data, m = 10, B = 5000, cl = 1, type = "lm")
}
\arguments{
\item{formula}{an object with class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame or a vector of filenames}

\item{m}{the dataset is split into m files.}

\item{B}{bootstrap iteration times}

\item{cl}{number of cores we use in parallel computing}

\item{type}{type of the model, either lm for linear regression or glm for logisic regression}
}
\value{
blblm object
}
\description{
main function to run linear regression with mini bag bootstrap
}
