% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{obtain coefficients}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{blblm object}

\item{...}{dditional arguments to be passed to the low level regression fitting functions}
}
\description{
obtain coefficients
}
