% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{col_types}
\alias{col_types}
\title{Show Column Types}
\usage{
col_types(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A data frame of names and types.
}
\description{
Gives a list of column names and types.
}
