% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{obtain confidence interval}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{an object with class "blblm}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{user specified confidence level, with default value 95\%}

\item{...}{dditional arguments to be passed to the low level regression fitting functions}
}
\description{
obtain confidence interval
}
