% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovidTools.R
\docType{package}
\name{CovidTools}
\alias{CovidTools}
\title{CovidTools: A package for visualizing Covid data for Indiana. Used as a final project at IUPUI.}
\description{
The CovidTools package provides the following visualization functions:#'
}
\section{functions}{

plot_deaths, plot_tests, plot_positiveTests, plot_positivityOfTest, plot_countyDeaths, plot_death_difference
}

