% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean.R
\name{dataclean}
\alias{dataclean}
\title{dataclean}
\usage{
dataclean(data)
}
\arguments{
\item{data}{the imported rnaseq dataset (excel file)}
}
\description{
Takes in an rnaseq dataset and cleans it so that gene expression data can be analyzed in subsequent functions.
}
\examples{
dataclean()
}
\keyword{clean,}
\keyword{dataset}
