% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZfilter.r
\name{EZfilter}
\alias{EZfilter}
\title{EZ Filter}
\usage{
EZfilter(dataFrame, label, covar = 0.5)
}
\arguments{
\item{dataFrame}{name of data frame}

\item{label}{unique column name to assign to coefficient of variance}

\item{covar}{coefficient of variance with default value 0.5}
}
\description{
Identifies coefficient of variance or degree of variability among samples for a given gene in a large data set and adds a new column for the coefficient of variance to the existing data frame. The "COVAR" column indicates the number of samples with coefficient of variance greater than a given criteria. The default criteria is set at 0.5. For example, if coefficient of variance for samples 1-5 for gene 1 is greater than 50\%, the "COVAR" column reads 5 meaning all 5 samples have coefficient of variance of greater than 50\% for gene 1.
}
\examples{
df <- EZfilter(df, "SAMPLE1")
}
\keyword{EZ}
\keyword{coefficient,}
\keyword{filter,}
\keyword{variance,}
