% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZfoldChange.r
\name{EZfoldChange}
\alias{EZfoldChange}
\title{EZ Fold Change}
\usage{
EZfoldChange(dataFrame, label)
}
\arguments{
\item{dataFrame}{name of data frame}

\item{label}{unique column name previously provided to EZmeanAndstd function}
}
\description{
Computes log2 fold change for each sample in a large data set and adds a new column for the calculated value for each sample to the existing data frame.
}
\examples{
df <- EZfoldChange(df, "SAMPLE1")
}
\keyword{EZ}
\keyword{change,}
\keyword{fold}
\keyword{foldchange,}
