% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZmeanAndStd.r
\name{EZmeanAndstd}
\alias{EZmeanAndstd}
\title{EZ Mean & Standard Deviation}
\usage{
EZmeanAndstd(dataFrame, columns, label)
}
\arguments{
\item{dataFrame}{name of data frame}

\item{columns}{a vector of columns to calculate mean and standard deviation}

\item{label}{unique column name to assign to the mean and standard deviation column}
}
\description{
Computes mean and standard deviation for each sample set in large data set and adds a new column for mean and std calculated for each sample set to the existing data frame. This functions requires tidyverse and matrixstats packages.
}
\examples{
df <- EZmeanAndstd(df, c('SAMPLE.1', 'SAMPLE.1.1', 'SAMPLE.1.2'), "SAMPLE1")
}
\keyword{EZ}
\keyword{mean,}
\keyword{std,}
