% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frpo_plot.R
\name{frpo_plot}
\alias{frpo_plot}
\title{Plots the frequency of SNPs over the length of the genome.}
\usage{
frpo_plot(x, total = 4000)
}
\arguments{
\item{total}{The length of the genome.}

\item{a}{A data.frame containing the data}
}
\value{
Graph of SNP frequency over genome
}
\description{
Plots the frequency of SNPs over the length of the genome.
}
\examples{
frpo_plot(a, 2000)
frpo_plot(a, 3500)
}
