% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{gibbs_sampling}
\alias{gibbs_sampling}
\title{Gibbs sampling function}
\usage{
gibbs_sampling(S = 1000, Ya, Yal, K = 20, nu_0 = 10, sigma2_0 = 6)
}
\arguments{
\item{S}{The number of iterations (default value is 1000)}

\item{Ya}{(M * N) matrix containing count data on all internal nodes of binary tree (M) and N is the number of samples (tree)}

\item{Yal}{(M * N) matrix containing count data on all left child of internal nodes of binary tree (M)}

\item{K}{Predefined number of clusters. This value is usually pretty large and will automatically decrease to the correct number of clusters (default value is 20)}

\item{nu_0}{The prior distribution of diagonal variance term is gamma and nu_0 / 2 is the shape parameter of gamma distribution (default value is 10)}

\item{sigma2_0}{(nu_0 * sigma2_0) / 2 is the scale parameter of gamma distribution (default value is 6)}
}
\value{
gibbs_sampling function returns a list containing one matrix (N * S) which contains posterior samples of label
}
\description{
Gibbs sampling function
}
\examples{
gibbs_sampling(S = 1000, Ya = matrix(c(10, 20, 25, 8, 12, 20), nrow = 2, byrow = TRUE), Yal = matrix(c(8, 12, 20, 4, 6, 10), nrow = 2, byrow = TRUE), K = 20, nu_0 = 10, sigma2_0 = 6)
}
