% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_boot}
\alias{glm_each_boot}
\title{logistic regression on one bootstrap}
\usage{
glm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{data frame}

\item{n}{total sample size}
}
\description{
logistic regression on one bootstrap
}
