% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{replicate glm fitting}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{a formula class object}

\item{data}{data frame for analysis}

\item{n}{sample size}

\item{B}{bootstrap iteration times}
}
\description{
replicate glm fitting
}
