% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautify.R
\name{\%or\%}
\alias{\%or\%}
\title{Apply Multiple Highlights}
\usage{
a \%or\% b
}
\arguments{
\item{data}{A data frame.}

\item{what}{What to highlight: "na", "max", "min".}

\item{colset}{A list of colors.}
}
\value{
A datatable.
}
\description{
Applies NA, max, or min highlights in one step.
}
