% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_max.R
\name{highlight_max}
\alias{highlight_max}
\title{Highlight Maximum Values}
\usage{
highlight_max(df, color = "lightgreen")
}
\arguments{
\item{df}{A dataframe.}

\item{color}{Background color for maximum values. Default is 'lightgreen'.}
}
\value{
A \code{DT::datatable} object with max values highlighted.

A DT datatable object with maximum values highlighted.
}
\description{
Highlights the maximum values in each numeric column of a data frame.
}
\examples{
df <- data.frame(a = c(1, 2, 3), b = c(5, 9, 6))
highlight_max(df)
Highlight Maximum Values in Each Numeric Column

}
