% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_min.R
\name{highlight_min}
\alias{highlight_min}
\title{Highlight Minimum Values}
\usage{
highlight_min(df, color = "lightblue")
}
\arguments{
\item{df}{A dataframe.}

\item{color}{Background color for minimum values. Default is 'lightblue'.}
}
\value{
A \code{DT::datatable} object with min values highlighted.

A DT datatable object with minimum values highlighted.
}
\description{
Highlights the minimum values in each numeric column of a data frame.
}
\examples{
df <- data.frame(a = c(1, 2, 3), b = c(5, 9, 6))
highlight_min(df)
Highlight Minimum Values in Each Numeric Column

}
