% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_na.R
\name{highlight_na}
\alias{highlight_na}
\title{Highlight Missing Values (NA)}
\usage{
highlight_na(df, color = "lightpink")
}
\arguments{
\item{df}{A dataframe.}

\item{color}{Background color for NA cells. Default is 'lightpink'.}
}
\value{
A \code{DT::datatable} object with NA values highlighted.

A DT datatable object with NA cells highlighted.
}
\description{
Highlights all NA values in a data frame by applying a background color using the DT package.
}
\examples{
df <- data.frame(a = c(1, NA, 3), b = c(NA, 2, 3))
highlight_na(df)
Highlight NA values in a DataFrame

}
