% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansplotter.R
\name{kmeansplotter}
\alias{kmeansplotter}
\title{kmeansplotter}
\usage{
kmeansplotter(finalexp)
}
\arguments{
\item{finalexp}{the cleaned rnaseq dataset}
}
\description{
Makes a k-means plot based on the TCA analysis run in tcaplot.R.
}
\examples{
kmeansplotter()
}
\keyword{analysis}
\keyword{expression}
\keyword{k-means}
\keyword{plot,}
