% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{fit regression model}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{an data frame}

\item{freqs}{freqency of each observations}
}
\description{
fit regression model
}
