% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{linear regression on one bootstrap}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{data frame}

\item{n}{total sample size}
}
\description{
linear regression on one bootstrap
}
