% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikehood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate Log-Likelihood for Logistic Regression}
\usage{
log_likelihood(beta, X, y)
}
\arguments{
\item{beta}{Numeric vector of coefficients.}

\item{X}{Numeric matrix of predictors.}

\item{y}{Numeric vector of binary response values (0 or 1).}
}
\value{
Negative log-likelihood value (as optim minimizes by default).
}
\description{
This function calculates the log-likelihood for a logistic regression model given the coefficients.
}
