% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midhinge.R
\name{midhinge}
\alias{midhinge}
\title{Takes the averages of q1 and q3.}
\usage{
midhinge(num, num_word, item, verb, adjective, location)
}
\arguments{
\item{num}{An integer}

\item{num_word}{A string corresponding to the integer}

\item{item}{A string}

\item{verb}{A string}

\item{adjective}{A string}

\item{location}{A string}
}
\value{
A string containing the words in grammatical order.
}
\description{
Takes a noun and makes it plural
}
