% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_SNP.R
\name{map_SNP}
\alias{map_SNP}
\title{Plots the number of SNPs observed at each passage number over time}
\usage{
map_SNP(x)
}
\arguments{
\item{a}{A data.frame containing the data}
}
\value{
Graph of number of SNPS over passage number
}
\description{
Plots the number of SNPs observed at each passage number over time
}
\examples{
map_SNP(a)
}
