% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_countyDeaths}
\alias{plot_countyDeaths}
\title{Creates a map of deaths by county due to Covid since March}
\usage{
plot_countyDeaths()
}
\description{
Creates a map of deaths by county due to Covid since March
}
\examples{
plot_countyDeaths()
}
