% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_deaths}
\alias{plot_deaths}
\title{Show plot of Deaths from Covid in Indiana since March}
\usage{
plot_deaths()
}
\description{
Show plot of Deaths from Covid in Indiana since March
}
\examples{
plot_deaths()
}
