% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_positiveTests}
\alias{plot_positiveTests}
\title{Show plot of number of positive tests performed in Indiana since March}
\usage{
plot_positiveTests()
}
\description{
Show plot of number of positive tests performed in Indiana since March
}
\examples{
plot_positiveTests()
}
