% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_positivityOfTest}
\alias{plot_positivityOfTest}
\title{Plots positivity rate of tests that have been performed since March}
\usage{
plot_positivityOfTest()
}
\description{
Plots positivity rate of tests that have been performed since March
}
\examples{
plot_positivityOfTest()
}
