% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_tests}
\alias{plot_tests}
\title{Show plot of number of tests performed in Indiana since March}
\usage{
plot_tests()
}
\description{
Show plot of number of tests performed in Indiana since March
}
\examples{
plot_tests()
}
