% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{prediciton for blbm class}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class "blblm"}

\item{new_data}{new data frame for prejection}

\item{confidence}{draw CI or not}

\item{level}{user specified confidence level, with default value 95\%}

\item{...}{additional arguments to be passed to the low level regression fitting functions}
}
\description{
prediciton for blbm class
}
