% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_logisticreg_app}
\alias{run_logisticreg_app}
\title{Run the Logistic Regression Shiny App}
\usage{
run_logisticreg_app()
}
\description{
This function launches the Shiny app for logistic regression.
The app allows users to upload datasets, select features,
and compute logistic regression results.
}
