% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGadget.R
\name{selectData}
\alias{selectData}
\title{A shiny gadget to select data from the timeline of Indiana deaths due to covid since march}
\usage{
selectData(data, xvar, yvar)
}
\description{
A shiny gadget to select data from the timeline of Indiana deaths due to covid since march
}
\examples{
selectData()
}
