% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{obtain sigma}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class "blblm"}

\item{confidence}{need confidence interval or not}

\item{level}{the confidence level required.}

\item{...}{dditional arguments to be passed to the low level regression fitting functions}
}
\description{
obtain sigma
}
