% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{This function take a data frame and output a list of data frame.
The list has length m. Each element in the list has the same dimension
as the original data but they could be repeated because they are
formed through bootstrap resampling.}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{a data frame}

\item{m}{the dataset is split into m files.}
}
\description{
This function take a data frame and output a list of data frame.
The list has length m. Each element in the list has the same dimension
as the original data but they could be repeated because they are
formed through bootstrap resampling.
}
