% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxvals.R
\name{tag_high}
\alias{tag_high}
\title{Highlight Max Values}
\usage{
tag_high(d, shade = "lightgreen")
}
\arguments{
\item{d}{A data frame.}

\item{shade}{Color to highlight the max. Default is lightgreen.}
}
\value{
A datatable.
}
\description{
Marks the largest number in each numeric column.
}
