% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minvals.R
\name{tag_low}
\alias{tag_low}
\title{Highlight Min Values}
\usage{
tag_low(d, shade = "lightblue")
}
\arguments{
\item{d}{A data frame.}

\item{shade}{Color to highlight the min. Default is lightblue.}
}
\value{
A datatable.
}
\description{
Marks the smallest number in each numeric column.
}
