% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{tag_na}
\alias{tag_na}
\title{Highlight NA Values}
\usage{
tag_na(d, shade = "lightpink")
}
\arguments{
\item{d}{A data frame.}

\item{shade}{Color for NA cells. Default is lightpink.}
}
\value{
A datatable.
}
\description{
Highlights all NA values in a data frame.
}
