% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcaplot.R
\name{tcaplot}
\alias{tcaplot}
\title{tcaplot}
\usage{
tcaplot(geneexp, gname)
}
\arguments{
\item{gname}{the chosen gene expression data for analysis}

\item{finalexp}{the cleaned rnaseq dataset}
}
\description{
Makes a TCA plot for a chosen gene from the cleaned dataset produced using dataclean.R.
}
\examples{
tcaplot()
}
\keyword{TCA}
\keyword{analysis}
\keyword{expression}
\keyword{plot,}
