% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_mean.R
\name{trim_mean}
\alias{trim_mean}
\title{Cuts the edges off of data to give more specialized mean}
\usage{
pluralize_gift(gift)
}
\arguments{
\item{gift}{A string or vector of strings}
}
\value{
A string or vector of strings with the pluralized words
}
\description{
Takes a noun and makes it plural
}
