% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_functions.R
\name{plot_death_difference}
\alias{plot_death_difference}
\title{Show plot of Deaths from Covid in Indiana since March compared to
the deaths we expected to see in 2020 prior to Covid}
\usage{
plot_death_difference()
}
\description{
Show plot of Deaths from Covid in Indiana since March compared to
the deaths we expected to see in 2020 prior to Covid
}
\examples{
plot_death_difference()
}
